mybatis generator 使用main()方法自动更新配置文件并运行

+++ zheng-cms/zheng-cms-dao/src/main/java/com/zheng/cms/dao/mapper/CmsArticleCategoryMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface CmsArticleCategoryMapper {
-    int countByExample(CmsArticleCategoryExample example);
+    long countByExample(CmsArticleCategoryExample example);
 
     int deleteByExample(CmsArticleCategoryExample example);
