+++ zheng-cms/zheng-cms-dao/src/main/java/com/zheng/cms/dao/mapper/CmsArticleTagMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface CmsArticleTagMapper {
-    int countByExample(CmsArticleTagExample example);
+    long countByExample(CmsArticleTagExample example);
 
     int deleteByExample(CmsArticleTagExample example);
